refactor(00-Project): refactor price to big decimal

+++ 00-PROJECT/MiniMart/backend/src/main/java/com/backend/common/dto/ProductLowStockDTO.java
@@ -1,5 +1,7 @@
 package com.backend.common.dto;
 
+import java.math.BigDecimal;
+
 import com.backend.product.model.ProductStatus;
 
 import lombok.AllArgsConstructor;
@@ -22,9 +24,9 @@ public class ProductLowStockDTO {
 
     private ProductStatus status;
 
-    private Double salePrice;
+    private BigDecimal salePrice;
 
-    private double compareAtPrice;
+    private BigDecimal compareAtPrice;
 
     @Getter
     @Setter
